// src/lib/types.ts
export type Skills =
  | "AWS"
  | "CSS"
  | "Docker"
  | "Figma"
  | "Git"
  | "HTML"
  | "JavaScript"
  | "Linux"
  | "Mocha"
  | "MongoDB"
  | "MySQL"
  | "Next.js"
  | "Node.js"
  | "PHP"
  | "PostgreSQL"
  | "Python"
  | "React"
  | "Redis"
  | "SanityCMS"
  | "Sass"
  | "Selenium"
  | "Tailwind CSS"
  | "TypeScript"